decouple context from host_function
