+++ nucleus/cage/src/host_func/kvdb.rs
@@ -1,8 +1,8 @@
-use super::*;
+use crate::mem;
 use codec::Encode;
 use rocksdb::{ColumnFamilyDescriptor, Options, DB};
 use vrs_core_sdk::{error::RuntimeError, CallResult, BUFFER_LEN, NO_MORE_DATA};
-use wasmtime::{Caller, FuncType, Val, ValType};
+use wasmtime::{Caller, Engine, FuncType, Val, ValType};
 
 pub(crate) fn init_rocksdb(path: impl AsRef<std::path::Path>) -> anyhow::Result<DB> {
     let avs_cf = ColumnFamilyDescriptor::new("avs", Options::default());
@@ -56,8 +56,8 @@ pub(crate) fn storage_put_signature(engine: &Engine) -> FuncType {
 /// ```
 /// fn storage_put(key_ptr: *const u8, key_len: i32, value_ptr: *const u8, value_len: i32, return_ptr: *mut u8) -> i32;
 /// ```
-pub(crate) fn storage_put(
-    mut caller: Caller<'_, Context>,
+pub(crate) fn storage_put<R>(
+    mut caller: Caller<'_, R>,
     params: &[Val],
     result: &mut [Val],
 ) -> anyhow::Result<()> {
@@ -67,7 +67,7 @@ pub(crate) fn storage_put(
         let return_value = CallResult::<()>::Err(RuntimeError::WriteIsNotAllowInGetMethod);
         let bytes = return_value.encode();
         assert!(bytes.len() <= BUFFER_LEN);
-        Context::write_bytes_to_memory(&mut caller, r_ptr, &bytes).expect("write to wasm failed");
+        mem::write_bytes_to_memory(&mut caller, r_ptr, &bytes).expect("write to wasm failed");
         return Ok(());
     }
     let k_ptr = params[0].unwrap_i32();
@@ -75,9 +75,9 @@ pub(crate) fn storage_put(
     let v_ptr = params[2].unwrap_i32();
     let v_len = params[3].unwrap_i32();
     let key =
-        Context::read_bytes_from_memory(&mut caller, k_ptr, k_len).expect("read from wasm failed");
+        mem::read_bytes_from_memory(&mut caller, k_ptr, k_len).expect("read from wasm failed");
     let val =
-        Context::read_bytes_from_memory(&mut caller, v_ptr, v_len).expect("read from wasm failed");
+        mem::read_bytes_from_memory(&mut caller, v_ptr, v_len).expect("read from wasm failed");
     let return_value = if let Err(e) = db_put(&caller.data().db, &key, &val) {
         CallResult::<()>::Err(RuntimeError::KvStorageError(e))
     } else {
@@ -85,7 +85,7 @@ pub(crate) fn storage_put(
     };
     let bytes = return_value.encode();
     assert!(bytes.len() <= BUFFER_LEN);
-    Context::write_bytes_to_memory(&mut caller, r_ptr, &bytes).expect("write to wasm failed");
+    mem::write_bytes_to_memory(&mut caller, r_ptr, &bytes).expect("write to wasm failed");
     Ok(())
 }
 
@@ -115,8 +115,8 @@ pub(crate) fn storage_get_signature(engine: &Engine) -> FuncType {
 ///         bytes: max 64k-4
 /// rest:
 ///   bytes: max 64k
-pub fn storage_get(
-    mut caller: Caller<'_, Context>,
+pub fn storage_get<R>(
+    mut caller: Caller<'_, R>,
     params: &[Val],
     result: &mut [Val],
 ) -> anyhow::Result<()> {
@@ -124,14 +124,14 @@ pub fn storage_get(
     let k_len = params[1].unwrap_i32();
     let r_ptr = params[2].unwrap_i32();
     let v_offset = params[3].unwrap_i32();
-    let key = Context::read_bytes_from_memory(&mut caller, k_ptr, k_len)
-        .expect("can't read bytes from wasm");
+    let key =
+        mem::read_bytes_from_memory(&mut caller, k_ptr, k_len).expect("can't read bytes from wasm");
     let db = &caller.data().db;
     let r = match db_get(db, &key) {
         Ok(value) => CallResult::<Option<Vec<u8>>>::Ok(value),
         Err(e) => CallResult::<Option<Vec<u8>>>::Err(RuntimeError::KvStorageError(e)),
     };
-    let flag = Context::write_to_memory(&mut caller, r_ptr, r, Some(v_offset))?;
+    let flag = mem::write_to_memory(&mut caller, r_ptr, r, Some(v_offset))?;
     result[0] = flag;
     Ok(())
 }
@@ -154,8 +154,8 @@ pub(crate) fn storage_del_signature(engine: &Engine) -> FuncType {
 /// ```
 /// fn storage_del(key_ptr: *const u8, key_len: i32, return_ptr: *mut u8) -> i32;
 /// ```
-pub fn storage_delete(
-    mut caller: Caller<'_, Context>,
+pub fn storage_delete<R>(
+    mut caller: Caller<'_, R>,
     params: &[Val],
     result: &mut [Val],
 ) -> anyhow::Result<()> {
@@ -165,13 +165,13 @@ pub fn storage_delete(
         let return_value = CallResult::<()>::Err(RuntimeError::WriteIsNotAllowInGetMethod);
         let bytes = return_value.encode();
         assert!(bytes.len() <= BUFFER_LEN);
-        Context::write_bytes_to_memory(&mut caller, r_ptr, &bytes).expect("write to wasm failed");
+        mem::write_bytes_to_memory(&mut caller, r_ptr, &bytes).expect("write to wasm failed");
         return Ok(());
     }
     let k_ptr = params[0].unwrap_i32();
     let k_len = params[1].unwrap_i32();
-    let key = Context::read_bytes_from_memory(&mut caller, k_ptr, k_len)
-        .expect("can't read bytes from wasm");
+    let key =
+        mem::read_bytes_from_memory(&mut caller, k_ptr, k_len).expect("can't read bytes from wasm");
     let return_value = if let Err(e) = db_del(&caller.data().db, &key) {
         CallResult::<()>::Err(RuntimeError::KvStorageError(e))
     } else {
@@ -179,6 +179,6 @@ pub fn storage_delete(
     };
     let bytes = return_value.encode();
     assert!(bytes.len() <= BUFFER_LEN);
-    Context::write_bytes_to_memory(&mut caller, r_ptr, &bytes).expect("write to wasm failed");
+    mem::write_bytes_to_memory(&mut caller, r_ptr, &bytes).expect("write to wasm failed");
     Ok(())
 }
