+++ nucleus/core-sdk/src/lib.rs
@@ -8,9 +8,6 @@ pub mod timer;
 pub use codec;
 pub use constant::*;
 pub use paste::paste as macro_paste;
-// TODO remove this
-// pub use vrs_primitives::{AccountId, Balance, BlockNumber, Hash, Nonce, Signature};
-
 pub use timer::_set_timer;
 
 /// the buffer used for transfering data from host to wasm
