+++ nucleus/test-suite/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "vrs-test-suite"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
