+++ nucleus/test-suite/src/lib.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("Hello, world!");
+}
